more change

+++ forms/book.go
@@ -5,5 +5,5 @@ type BookForm struct {
 	Title       string `form:"title" json:"title" binding:"required,max=100"`
 	Autor       string `form:"autor" json:"autor" binding:"required,max=100"`
 	Description string `form:"desctiption" json:"desctiption" binding:"required,max=1000"`
-	LibraryID   int    `db:"library_id, foreignkey" json:"library_id"`
+	LibraryID   int64  `db:"library_id, foreignkey" json:"library_id"`
 }
